feat: Transfer LSTM/RNN code from notebook to files. Move text handling to seperate file.
The code from the Jupyter Notebook is transferred to a file for easier handling and automatization.
All code for the text handling was put together into one class and this one moved to
a single file to allow access from all models.
So far there has not been much testing, only restructuring the code, but we need to checkpoint
this stage.
Unit tests must be extended, so far only renaming issues done so that tests are all green.